requirements.txt
# requirements.txt file is a file that lists all of the modules needed for the 
# Django project to work. 


asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
gunicorn==20.1.0
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.6.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
python-dotenv==0.19.0